Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float num1, num2; 
-    float div, prod, add, minus;
+    int num1, num2; 
+    int div, prod, add, minus;
     char ch;
     scanf("%d %d %c",&num1, &num2, &ch);
     add = num1+num2;
